Renamed Input folder again
